backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pandas>=2.0.0
numpy>=1.24.0
groq>=0.9.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openpyxl>=3.1.0
pdfplumber>=0.10.0
pydantic>=2.0.0
